Added more game commands.
The game is roughly playable as jack using half the full map. (I still need to enter the rest of the coords for the lower half.)

Jack doesn't yet use any transportation cards.

Also, investigators cannot yet search for clues or make an arrest, so the game is a bit one-sided. :-D