Minor changes
Jack is less likely to vacillate between two destinations when getting cornered.  He now uses the unweighted move cost to choose the target, but then uses the weighted move cost to choose the path to take to get there.

Also updated game start to allow the investigator positions to be defined before Jack begins the game.

Made the graph coloring a callable function in preparation for a future change to redraw the graph with discovery, clue, and current investigator locations each turn.