More Jack behavior tweaking
Weights spaces 185, 112, and 97 that same as water paths since they are risky bottlenecks and choosing a boat is usually a better option if you must traverse that area of the map.

Also added logic for Jack to use a coach if an investigator is 1 space away.

Finally, save the map PDF file after every move so it can be used as a reference showing clue, discovery, and investigator positions as the game progresses.