Bug fixes
If there were coach cards left and Jack computed a path using a special move (such as boat or alley) but didn't have any cards left of that type, it would detect a coach needed to be used, but then would use the boat or alley.  If has now been fixed to have the coach move preempt all others when the cost is > 1000 (which implies Jack does not have the needed card).